:zap: :cop: Separate REPL part
